requirements.txt
# KidOS Agentic AI - MVP Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0
chromadb>=0.4.0
sse-starlette>=1.6.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
